Code review changes - Move SiddhiAppRuntime to processor core
