+++ components/org.wso2.carbon.stream.processor.core/pom.xml
@@ -218,7 +218,7 @@
             osgi.service; objectClass="org.wso2.msf4j.Microservice"; serviceCount="2",
             osgi.service; objectClass="javax.ws.rs.ext.ExceptionMapper";serviceCount="5",
             osgi.service; objectClass="org.wso2.carbon.deployment.engine.Deployer",
-            osgi.service; objectClass="org.wso2.carbon.stream.processor.common.SiddhiAppRuntimeService"
+            osgi.service; objectClass="org.wso2.carbon.stream.processor.core.SiddhiAppRuntimeService"
         </carbon.component>
     </properties>
